code/requirements.txt
tensorflow>=2.12
datasets>=2.14.0
transformers>=4.36.0
numpy>=1.23.0
Pillow>=9.5.0
PyYAML>=6.0
tqdm>=4.66.0
